Title: tuning performance of MQTT and TCP socket API
Comment:
1.  The available takes too long to check packets. Temporal change recv timeout to 10s for peek packet.
2.  TCP read 1 character timeout cause connection disconnect. Add checking return value to avoid this
3.  MQTT use busy loop if no coming data

+++ hardware/cores/arduino/server_drv.h
@@ -41,7 +41,7 @@ public:
     /* Usually used by UDP */
 	bool sendtoData(int sock, const uint8_t *data, uint16_t len, uint32_t peer_ip, uint16_t peer_port);
 
-	uint16_t availData(int sock);
+	int availData(int sock);
 	
 	void getRemoteData(int sock, uint32_t *ip, uint16_t *port);
